Move the REGISTRATION_OPEN logic into the base view.
Since both of the included backends already use the
``REGISTRATION_OPEN`` setting to determine whether to allow
registration, and that approach works even when ``REGISTRATION_OPEN``
isn't explicitly set, there's no reason to require both backends to
explicitly implement the logic.

This is backwards-compatible with the prior implementation, so long as
``REGISTRATION_OPEN`` was not set.