Handle quirks of HMAC + username regexes.
This cleans up the default URL pattern of the HMAC activation view,
basing it off Django's own default username regex, expanded to handle
the extra characters which can appear in the HMAC-signed
value. Additionally, a warning is added to the documentation about the
need to ensure that the URL pattern for the HMAC activation view will
permit the full set of potential legal characters.