Refactor the HMAC use a bit.
Per Florian's suggestion, the HMAC workflow now uses signing.dumps()
and signing.loads() to generate and validate the activation key. This
slightly changes the necessary URL pattern, but is also a bit cleaner
because it avoids potential character-encoding issues from having the
raw username be in the pattern (dumps() will base64-encode, so only
URL-safe base64 characters can be produced).