Remove registration from INSTALLED_APPS in tests.
Since only the model-activation backend actually needs registration in
INSTALLED_APPS, tests of the other backends, and of the forms, now run
without that. Unfortunately it's not (as far as I can tell) possible
to default to registration not being in INSTALLED_APPS and add it back
only for model-activation tests, due to when the migrations are
applied.