Use custom user model in RegistrationForm
By default, Django will use its own default User model when
initializing the CreationForm.

This will cause FieldErrors to be raised when a customer user model
is provided but doesn't contain the same fields as in Django's default
implementation.

That's why we use the custom user model to instantiate a CreationForm
instead. The implementation can now have any fields it wants as long as
it's compatible with Django's AbstractUser.

Fixes #234